Flat: Reset the bt module before hciattach
On the RevPi Flat the bt module is connected to a USB UART. This creates
some problems. Currently the kernel is not able to use serdev for the
USB UART. Which in return makes it impossible to define the bluetooth
device in the devicetree. The bt driver has a function to handle the
BT_REG_ON pin of the bt module. But this is not possible without the
devicetree.

Without a proper reset of the bt module the module doesn't respond to
the hciattach. As the module is not reset during a reboot, we need to
reset the module before we attach the module.

We toggle the BT_REG_ON pin to reset the module. The pin is connected to
the GPIO11 on the GPIO expander on the top board. The BT power-on reset
(POR) circuit is out of reset after BT_REG_ON goes High.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>