Add check that the rules are only run on Flat
The revpi-bluetooth package is currently only needed for Revpi Flat. But
it will be part of the normal image. So we need to make sure the rules
are only running on RevPi Flat. As the bluetooth UART is connected via
USB on any other device a USB UART might get unumerated the same. So the
matching on the rules might not be enough to enforce this.

The rules match now also on the device compatible: kunbus,revpi-flat.
This ensures the rules are only run on the correct device.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>