Switch back to the debian/ namespace
Raspberry Pi OS is reverting more and more back to Debian with
extensions. We introduced the raspios/ name space to be able to create
packages for RaspiOS and Debian. But with RaspiOS just being an
extension to Debian this is not needed anymore.

Replace the raspios/ name space with the debian/ namespace starting with
the Bookworm release.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>