fix(btuart): catch actual PID from hciattach
While the "-p" flag of "hciattach" does indeed print the PID of the
forked "hciattach" process, it doesn't *just* print the PID. This made
the call to "kill" fail as stdout printed text as well as the PID.

Fix this by not letting "hciattach" fork to the background and start it
asynchronously in the shell, using the job's PID to immediately kill the
job. This is more robust than parsing the output of "hciattach".

Fixes: f6efaa2 ("feat: Add support for Flat S")
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>