fix(btuart): give Flat S bt module time to init
The bluetooth module on the Flat S seems to require a little bit more
time to initialise than it should. On initialisation, the kernel seems
to send the opcode 0x1003 to the module, which fails to answer in a
timely manner, causing a timeout:

    [   56.232153] Bluetooth: hci0: command 0x1003 tx timeout
    [   56.232168] Bluetooth: hci0: Opcode 0x1003 failed: -110

The second time the module is initialised there is no timeout.

It's unknown why exactly this happens, but it's assumed that either the
module takes too long to initialise, or the firmware is subtly broken.
Both aren't easily verifiable because the firmware source code isn't
published.

Previously the revpi-btuart script was supposed to solve this problem by
attaching the bluetooth module using the "hciattach" command when the
device is booted, using the udev rule in 90-revpi-bluetooth.rules, but
this didn't work reliably.

Give the bluetooth module some time to initialise between calls to
"hciattach" to make sure no timeout is reached. This also reliably makes
the bluetooth module available to the system after a successful boot.

Fixes: f6efaa2 ("feat: Add support for Flat S")
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>